Cleaner messages dispatching
The pattern matching over the messages was not the clearest part of
the code. Using the symbol names of the methods that we want to call
on the actors seems like more readable way to do so.

We still prefer pattern matching in the worker actor, as makes more
sense than the method-call-like approach.